refactor: 확장성을 고려하여 API 타입 정의 계층 분리 및 구조화 (#1)
