--- src/apis/api.ts
@@ -1,5 +1,0 @@
-const BASE_URL = '/api';
-
-export const api = {
-  savingsProducts: `${BASE_URL}/savings-products`,
-};
