+++ src/apis/index.ts
@@ -1,1 +1,1 @@
-export * from './api';
+export * from './savingsApi';
