+++ src/constants/baseUrl.ts
@@ -1,0 +1,1 @@
+export const API_BASE_URL = '/api';
