+++ src/types/apis/savingsApiTypes.ts
@@ -1,0 +1,10 @@
+export interface SavingsProductAPIResponse {
+  id: string;
+  name: string;
+  annualRate: number;
+  minMonthlyAmount: number;
+  maxMonthlyAmount: number;
+  availableTerms: number;
+}
+
+export type SavingsProductsAPIResponse = SavingsProductAPIResponse[];
